+++ cpu_demo/SVDImpl.hh
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <cstdint>
+#include <random>
 
 #include <eigen3/Eigen/Core>
 
@@ -17,17 +18,17 @@ class SVDImpl {
 public:
     SVDImpl(int nusers, int nmovies, int nfeatures);
 
-    void fit(uint32_t *users, uint16_t *movies, float *ratings,
+    void fit(uint32_t *users, uint32_t *movies, float *ratings,
                int npoints, float lrate, float reg, int niters, bool verbose);
 
-    float *predict(uint32_t *users, uint16_t *movies, int npoints);
+    float *predict(uint32_t *users, uint32_t *movies, int npoints);
 
 private:
 
     /**
      * @brief Make a prediction at the given (user, movie) pair.
      */
-    inline float predict(uint32_t user, uint16_t movie) const;
+    inline float predict(uint32_t user, uint32_t movie) const;
 
     /**
      * @brief As `predict`, but in-place (to avoid allocations) and on Eigen
@@ -37,7 +38,7 @@ private:
      */
     inline void predict_inplace(
             const Eigen::Array<uint32_t, 1, Eigen::Dynamic> &users,
-            const Eigen::Array<uint16_t, 1, Eigen::Dynamic> &movies,
+            const Eigen::Array<uint32_t, 1, Eigen::Dynamic> &movies,
             Eigen::Array<float, 1, Eigen::Dynamic> &result);
 
     int nfeatures;
